create documents app and related files and del files from tasks app
