+++ smse_onboarding_main_project/documents/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class DocumentsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'documents'
