+++ smse_onboarding_main_project/documents/templates/documents/list.html
@@ -1,0 +1,51 @@
+{% extends 'base.html' %}
+
+{% block content %}
+<div class="container mt-4">
+    <div class="card">
+        <div class="card-header d-flex justify-content-between align-items-center">
+            <h2>Documents</h2>
+            <button class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#uploadModal">
+                <i class="bi bi-upload"></i> Upload New Document
+            </button>
+        </div>
+        <div class="card-body">
+            <div class="table-responsive">
+                <table class="table table-hover">
+                    <thead>
+                        <tr>
+                            <th>Title</th>
+                            <th>Uploaded By</th>
+                            <th>Upload Date</th>
+                            <th>Actions</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        {% for doc in documents %}
+                        <tr>
+                            <td>{{ doc.title }}</td>
+                            <td>{{ doc.uploaded_by.username }}</td>
+                            <td>{{ doc.uploaded_at|date:"M d, Y" }}</td>
+                            <td>
+                                <a href="{% url 'documents:download' doc.document_id %}" class="btn btn-sm btn-info">
+                                    <i class="bi bi-download"></i> Download
+                                </a>
+                                <button class="btn btn-sm btn-danger" onclick="deleteDocument({{ doc.document_id }})">
+                                    <i class="bi bi-trash"></i> Delete
+                                </button>
+                            </td>
+                        </tr>
+                        {% empty %}
+                        <tr>
+                            <td colspan="4" class="text-center">No documents found.</td>
+                        </tr>
+                        {% endfor %}
+                    </tbody>
+                </table>
+            </div>
+        </div>
+    </div>
+</div>
+
+{% include 'documents/upload_modal.html' %}
+{% endblock %}
