+++ smse_onboarding_main_project/smse_onboarding/settings.py
@@ -41,6 +41,8 @@ INSTALLED_APPS = [
     'social_django',
     'tasks',
     'smse_onboarding',
+    #'tasks.apps.TasksConfig',
+    'documents.apps.DocumentsConfig',
 ]
 
 AUTHENTICATION_BACKENDS = [
