+++ smse_onboarding_main_project/tasks/models.py
@@ -111,47 +111,4 @@ class TaskProgress(models.Model):
         return f"{self.faculty} - {self.task} - Completed: {self.completed}"
 
 
-class FacultyDocument(models.Model):
-    """
-    Model for a document uploaded by a faculty user.
-    """
-    document_id = models.AutoField(primary_key=True)
-    faculty = models.ForeignKey(Faculty, on_delete=models.CASCADE, related_name='documents')
-    title = models.CharField(max_length=255)
-    file = models.FileField(upload_to='faculty_documents/')
-    uploaded_at = models.DateTimeField(auto_now_add=True)
-    uploaded_by = models.ForeignKey(User, on_delete=models.SET_NULL, null=True)
 
-    objects: Manager = Manager()
-
-    def __str__(self):
-        # Access the faculty instance properly
-        faculty_instance = self.faculty
-        return f"{self.title} - {faculty_instance.first_name}'s Document"
-
-    def delete(self, *args, **kwargs):
-        # Delete the file from storage
-        if self.file:
-            storage = self.file.storage
-            name = self.file.name
-            if storage.exists(name):
-                storage.delete(name)
-        super().delete(*args, **kwargs)
-
-
-class OtherEmployee(models.Model):
-    """
-    Model for other employees such as those from ITS and HR.
-    """
-    other_employee_id = models.AutoField(primary_key=True)
-    first_name = models.CharField(max_length=255)
-    last_name = models.CharField(max_length=255)
-    job_role = models.CharField(max_length=255)
-    other_employee_dept = models.CharField(max_length=255)
-    email = models.EmailField(max_length=255)
-    phone = models.CharField(max_length=10)
-    office_room = models.CharField(max_length=20)
-    associated_tasks = models.ManyToManyField(Task, blank=True)
-
-    def __str__(self):
-        return f"{self.first_name} {self.last_name}"
